Homologação V3.0 - Segunda Entrega (#8)
* feature(controller&service): CRUD User almost ready

* Feature: CRUD Almost Ready & Exceptions Treated

* Feature(ControllerAdvice): MethodArgumentNotValid almost treated

* Fix(UsuarioController): adding users

* Fix(User): Refactoring CRUD

* Feature(User-Controller): CPF is validated

* Docs: API Documentation Created V1.0

* Develop: Creating CRUD User and Address

* Test: CRUD User and Address

* Feature(User & Address): CRUD Created and Validated

* refactor: Removing flyway and Refactoring Database

* fix: POST User and Address Merged and Server Port changed

* Feature: New CORS added

* fix(handleExceptions): A better version for exceptions treated

* feature: Entities Modified and Services, new fields on the User

* refactor: Options Added at CORS and CrossOrigin removed

* Feature: GET User and Address for Query param

---------

+++ src/main/java/br/com/senai/sollaris/domain/resources/dtos/output/ReturnUsuarioDto.java
@@ -3,8 +3,8 @@ package br.com.senai.sollaris.domain.resources.dtos.output;
 import java.time.LocalDateTime;
 import java.util.List;
 
-import br.com.senai.sollaris.domain.Generos;
 import br.com.senai.sollaris.domain.Usuario;
+import br.com.senai.sollaris.domain.enums.Generos;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
@@ -16,7 +16,7 @@ import lombok.Setter;
 @NoArgsConstructor
 
 public class ReturnUsuarioDto {
-	private Long id;
+	private Integer id;
 	private LocalDateTime data_requisicao = LocalDateTime.now();
 	private String nome;
 	private Generos genero;
